Refactor AFI for 2D interpolation on either Re or UserProp
The module was refactored to manage data for a single airfoil file.  The calling code is now responsible or managing multiple airfoils by creating multiple AFI entities.
AFI was extended to support multiple Angle of Attack vs Cl,Cd, Cm tables which are dependent on either Reynolds number or a User Property.
This version requires the caller to specify whether 1D lookup on AoA, 2D on Re and AoA or 2D on UserProp and AoA is desired.
Future versions may support 3D lookup on AoA, Re, and UserProp.
The (Re or UserProp) linear interpolation is done first, and then the second (AoA) interpolation can be linear or spline-based.
A basic driver was added to test the interpolation scheme.