Reworks the KFC Driver
still needs to be tested (i am having issues with precision definition and NWTClib). Also changed a few input values. Updated the tether_force just before the controller is called. This brings the state to time t, instead of time t-1, and the problem was with t=0s, but should not affect anything major.